building_rag_project1/requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-openai
langchain-community
chromadb
pymupdf
pdfplumber
python-pptx
unstructured
python-multipart
pydantic
python-dotenv
